Extract AuthenticationWall to own file
